sched/fair: if sync flag ignored, try to place in same cluster
If the sync flag is ignored because the current cpu is not in
the affinity mask for the target of a sync wakeup (usually
binder call), prefer to place in the same cluster if
possible. The main case is a "top-app" task waking a
"foreground" task when the top-app task is running on
a CPU that is not in the foreground cpuset. This patch
causes the "top-app" search order be used (big cpus first)
even if schedtune.boost is 0 when the sync flag failed.

Based on: https://android.googlesource.com/kernel/msm.git/+/98ae57d9eaf7d12184ea339afd031083584092b4%5E%21/#F0

Made into a sched feature and added  tracepoints for testing

Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>